Refactoring webassembly cggmp test spec.
Trying to remove the template code generation but we have a problem
passing the key share back to playwright as ThresholdKeyShare contains
BTreeMap instances which are converted to Map by serde_wasm_bindgen
which does not play nice with with JSON stringify/parse.

Perhaps key shares will be encoded as JSON and then converted to PEM
for passing between webassembly and javascript.